requirements.txt
anthropic==0.50.0
discord.py==2.4.0
playwright==1.51.0
requests==2.32.3
beautifulsoup4==4.13.4
lxml==5.3.0
slack_sdk==3.35.0
APScheduler==3.11.0
tenacity==9.1.2
python-dotenv==1.1.0
